Consolidated Stops and Place (Place Type Extracted)
Pre-Processing Query 

The purpose of this query was to clean up and consolidate similar stop locations (like aerialway is = aeroway). This consolidation was done manually looking at the results from Query 3, and consolidating top stop locations & places was done so as to provide a truncated view of this information, making this data more readable and consistent when converting to images.