Consolidated Stops and Places (only tags)
Pre-Processing Query
The purpose of this query was to look at all the stop locations for all 4.45M trip instances, and provide a consolidated view of place type by looking at OpenStreetMap. The top level location recommendations in OpenStreetMap is based on the longitude and latitude of the stop and start vehicle movement data provided.